requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic[email]==2.10.4
python-dotenv==1.0.1
httpx==0.28.1

# OpenAI Agents SDK
openai-agents==0.0.7
openai==1.59.5

# Database
asyncpg==0.30.0
pgvector==0.3.6
alembic==1.14.0
sqlalchemy==2.0.36

# Kafka
aiokafka==0.12.0

# Channel integrations
google-api-python-client==2.157.0
google-auth-oauthlib==1.2.1
google-cloud-pubsub==2.27.1
twilio==9.4.0

# MCP Server
mcp==1.2.0

# Utilities
python-multipart==0.0.20
structlog==24.4.0
tenacity==9.0.0
textblob==0.18.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
locust==2.32.4
